feat: push to debug filter component
