+++ packages/client/src/schema-component/components/SchemaComponentProvider.tsx
@@ -1,11 +1,11 @@
-import React, { useMemo, useState } from 'react';
-import { uid } from '@formily/shared';
 import { createForm } from '@formily/core';
-import { useCookieState } from 'ahooks';
-import { useTranslation } from 'react-i18next';
 import { FormProvider } from '@formily/react';
-import { ISchemaComponentProvider } from '../types';
+import { uid } from '@formily/shared';
+import { useCookieState } from 'ahooks';
+import React, { useMemo, useState } from 'react';
+import { useTranslation } from 'react-i18next';
 import { SchemaComponentContext } from '../context';
+import type { ISchemaComponentProvider } from '../types';
 import { SchemaComponentOptions } from './SchemaComponentOptions';
 
 export const SchemaComponentProvider: React.FC<ISchemaComponentProvider> = (props) => {
