+++ packages/client/src/schema-component/index.tsx
@@ -1,6 +1,6 @@
 export * from './antd';
+export * from './antd/AntdSchemaComponentProvider';
 export * from './components';
 export * from './context';
 export * from './hooks';
 export * from './types';
-
